I want that on the home page, there will be two tab sender tab and receiver tab One user will upload the file on the server if he has uploaded the file on the server. When another user logged in, he should be directly on the receiver tab, and the file should be visible there directly with the time when it was uploaded and the user can download the file And user one should be able to delete his uploaded file, not others, but he and if he uploaded and then deleted the file, when user to login, he should be on the sender tab And both should see the status if another user is on send, logged in or logged out at the main purpose of this app is to
I have a website where I have to log in with my email and password, but so do my friend. So the solution we have found is, I will log in on the website, and that website provides a extension with the help of which we can download our cookies into a text file And then when we send the text file to our friend, he can input the text file into the extension, and can also login because in normal method, the website doesn't allow login from two devices at the same time. So we have found this method But the issue is, whenever I log in, And send cookies to my friend. And then when he export the cookies, his unable to luggage

 I want that when a user export a file, that file should be uploaded to the server . And when my friend open the extension should send an api call to our server, which we will deploy on render. You will have to help me with setting up the server and apply it on render, so that the aim is when I am logging in, I will export the file, and it will be sent to the server And when my friend want to log in, he will just open the extension and he will see the status there. If I am logged in, if I am logged in, he can import the file from the server directly when he input the file from the server, it will be downloaded in his pc. Then he can import it And if I am not online, he can just log in using password. This is so that I dont have to keep sending him file on Whatsapp regularly